pyproject.toml
[tool.poetry]
name = "movehub-backend"
version = "1.0.0"
description = "Production-grade backend for Uber-like moving companies marketplace"
authors = ["MoveHub Team"]
readme = "README.md"
python = "^3.11"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
pydantic = {extras = ["email"], version = "^2.5.0"}
pydantic-settings = "^2.1.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.25"}
asyncpg = "^0.29.0"
alembic = "^1.13.0"
geoalchemy2 = "^0.14.3"
shapely = "^2.0.2"

# Authentication & Security
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
cryptography = "^42.0.0"

# Redis & Caching
redis = {extras = ["hiredis"], version = "^5.0.1"}
fakeredis = {extras = ["lua"], version = "^2.21.0"}

# AWS Services
boto3 = "^1.34.0"
aioboto3 = "^12.3.0"

# OpenTelemetry - Comprehensive Observability
opentelemetry-api = "^1.22.0"
opentelemetry-sdk = "^1.22.0"
opentelemetry-instrumentation-fastapi = "^0.43b0"
opentelemetry-instrumentation-sqlalchemy = "^0.43b0"
opentelemetry-instrumentation-redis = "^0.43b0"
opentelemetry-instrumentation-logging = "^0.43b0"
opentelemetry-instrumentation-asyncpg = "^0.43b0"
opentelemetry-exporter-otlp = "^1.22.0"
opentelemetry-exporter-prometheus = "^0.43b0"
opentelemetry-instrumentation-aws-lambda = "^0.43b0"

# Payment Processing
stripe = "^7.11.0"

# Notifications
twilio = "^8.13.0"
sendgrid = "^6.11.0"

# Utilities
python-dateutil = "^2.8.2"
pytz = "^2023.3"
phonenumbers = "^8.13.27"
httpx = "^0.26.0"
tenacity = "^8.2.3"

# Validation & Serialization
email-validator = "^2.1.0"
orjson = "^3.9.12"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"
black = "^23.12.1"
ruff = "^0.1.11"
mypy = "^1.8.0"
pre-commit = "^3.6.0"
httpx = "^0.26.0"
faker = "^22.0.0"
factory-boy = "^3.3.0"

[tool.poetry.group.test.dependencies]
testcontainers = {extras = ["postgresql"], version = "^3.7.1"}
locust = "^2.20.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
    "redis.*",
    "geoalchemy2.*",
    "shapely.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--cov=app",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
    "--cov-fail-under=85",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "e2e: End-to-end tests",
    "slow: Slow running tests",
]

[tool.coverage.run]
source = ["app"]
omit = [
    "*/tests/*",
    "*/migrations/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
